refactor(tasks): move G1 joystick owner [skip ci]
