+++ src/unilab/envs/locomotion/__init__.py
@@ -1,3 +1,3 @@
-"""Locomotion env registry bootstrap contract."""
+"""Generic locomotion environment components."""
 
-__unilab_registry_modules__ = ("unilab.envs.locomotion.g1",)
+__unilab_registry_modules__: tuple[str, ...] = ()
