+++ src/unilab/envs/locomotion/g1/__init__.py
@@ -1,17 +1,1 @@
-from .joystick import (
-    G1WalkControlConfig,
-    G1WalkEnv,
-    G1WalkEnvCfg,
-    G1WalkFlatCfg,
-    G1WalkRewardConfig,
-    G1WalkRoughCfg,
-)
-
-__all__ = [
-    "G1WalkControlConfig",
-    "G1WalkEnv",
-    "G1WalkEnvCfg",
-    "G1WalkFlatCfg",
-    "G1WalkRewardConfig",
-    "G1WalkRoughCfg",
-]
+"""Legacy G1 shared base pending motion-tracking migration."""
