+++ src/unilab/tasks/__init__.py
@@ -11,7 +11,7 @@ __unilab_registry_modules__ = (
     "unilab.tasks.locomotion.go1",
     "unilab.tasks.locomotion.go2",
     "unilab.tasks.locomotion.go2w",
-    "unilab.envs.locomotion.g1",
+    "unilab.tasks.locomotion.g1",
     "unilab.tasks.locomotion.go2_arm",
     "unilab.tasks.locomotion.a2",
     "unilab.envs.manipulation.allegro_inhand",
