+++ src/unilab/tasks/locomotion/g1/__init__.py
@@ -1,0 +1,19 @@
+from .joystick import (
+    G1WalkControlConfig,
+    G1WalkEnv,
+    G1WalkEnvCfg,
+    G1WalkFlatCfg,
+    G1WalkRewardConfig,
+    G1WalkRoughCfg,
+)
+from .symmetry import G1SymmetryAugmentation
+
+__all__ = [
+    "G1SymmetryAugmentation",
+    "G1WalkControlConfig",
+    "G1WalkEnv",
+    "G1WalkEnvCfg",
+    "G1WalkFlatCfg",
+    "G1WalkRewardConfig",
+    "G1WalkRoughCfg",
+]
