+++ tests/envs/locomotion/g1/test_symmetry_contract.py
@@ -7,7 +7,7 @@ import torch
 
 from unilab.base import registry
 from unilab.base.registry import ensure_registries
-from unilab.envs.locomotion.g1.joystick import G1WalkRewardConfig
+from unilab.tasks.locomotion.g1.joystick import G1WalkRewardConfig
 
 pytest.importorskip("mujoco", reason="mujoco is required for G1 symmetry contract tests")
